#33 refactor(domain): 외부 통신용 ReservationId 값 객체 생성
