+++ payment/src/main/java/wisoft/nextframe/payment/domain/ReservationId.java
@@ -1,0 +1,23 @@
+package wisoft.nextframe.payment.domain;
+
+import java.util.Objects;
+import java.util.UUID;
+
+import lombok.Getter;
+
+@Getter
+public class ReservationId {
+	private final UUID value;
+
+	private ReservationId(UUID value) {
+		this.value = Objects.requireNonNull(value, "ReservationId는 null일 수 없습니다.");
+	}
+
+	public static ReservationId of(UUID value) {
+		return new ReservationId(value);
+	}
+
+	public static ReservationId of() {
+		return new ReservationId(UUID.randomUUID());
+	}
+}
