+++ payment/src/test/java/wisoft/nextframe/payment/fixture/TestPaymentFactory.java
@@ -4,8 +4,8 @@ import java.time.LocalDateTime;
 import java.util.UUID;
 
 import wisoft.nextframe.payment.common.Money;
+import wisoft.nextframe.payment.domain.ReservationId;
 import wisoft.nextframe.payment.domain.payment.Payment;
-import wisoft.nextframe.schedulereservationticketing.reservation.ReservationId;
 
 public class TestPaymentFactory {
