+++ payment/src/test/java/wisoft/nextframe/payment/infra/payment/persistence/PaymentEntityFixture.java
@@ -10,7 +10,7 @@ import wisoft.nextframe.payment.domain.payment.Payment;
 import wisoft.nextframe.payment.domain.payment.PaymentId;
 import wisoft.nextframe.payment.domain.payment.PaymentStatus;
 import wisoft.nextframe.payment.infra.payment.PaymentEntity;
-import wisoft.nextframe.schedulereservationticketing.reservation.ReservationId;
+import wisoft.nextframe.payment.domain.ReservationId;
 
 public class PaymentEntityFixture {
